moved api calls from search to bookgrid
