+++ src/Components/SearchBar.js
@@ -2,14 +2,12 @@ import { useState } from "react"
 import { Form } from "react-bootstrap";
 import { createSearchParams, useNavigate, useSearchParams} from "react-router-dom";
 
-export default function SearchBar({totalItems}){
+export default function SearchBar(){
   const [queryParams] = useSearchParams();
 
   const [pageNum, setPageNum] = useState(queryParams.get("page"));
   const [searchStr, setSearchStr] = useState(queryParams.get("search_bar"));
 
-
-  const navigate = useNavigate()
   return (
 
     <Form>
@@ -20,16 +18,7 @@ export default function SearchBar({totalItems}){
         onChange={(event) => setSearchStr(event.target.value)}
         />
       <button>Search</button>
-     
-      <input
-        name="page"
-        type="number"
-        value={pageNum}
-        onChange={(event) => setPageNum(event.target.value)}
-        min="1"
-        style={{width: "5rem"}}
-        />        
-      <button>Page Number:</button>
+
 
     </Form>
   );
